UI for Game Scene and Debug Menu for Cheat Codes
Includes functionality for exit, camera switching, enemy turns (3 second pause), and roll table display. The move button displays an empty scrollable viewport of where the moves transcript will be located. Also, the text for the stats is hooked up and resetting it works.

To access the debug menu, press escape key. Then you may type a command like "win", click out of the menu and press enter key, then the win screen is displayed. I made the debugger so that we can test screens that currently cannot be accessed.